Fix up tests, fix bug with multiple applicable extractors.

+++ src/TML.Tests/ExtractingTest.cs
@@ -1,58 +1,54 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using NUnit.Framework;
 using TML.Files;
-using TML.Files.Abstractions;
+using TML.Files.Extraction;
+using TML.Files.Extraction.Extractors;
 using TML.Patcher.Extractors;
-using TML.Patcher.Extractors.Info;
 
 namespace TML.Tests
 {
-    public class ExtractingTest
+    [TestFixture]
+    public static class ExtractingTest
     {
         private static readonly string[] Files =
         {
             "GamerMod.pdb",
             "GamerMod.dll",
             "build.txt",
-            "Info",
+            // "Info",
             "Properties/launchSettings.json",
             "description.txt",
             "icon.png",
             "workshop.json",
             "Gores/GoreTest.png",
-            "Gores/GoreTest.rawimg",
+            // "Gores/GoreTest.rawimg",
             "GamerMod.zip"
         };
-        
+
         [Test]
         public static void ContainsExactlyTheExpectedFiles() {
-            using Stream tmodFile = typeof(ExtractingTest).Assembly.GetManifestResourceStream("TML.Tests.GamerMod.tmod")!;
-            IModFileReader reader = new ModFileReader();
-            IModFileExtractor extractor = new ModFileExtractor();
-            IModFile file = reader.Read(tmodFile);
-            IFileExtractor[] extractors = {new InfoFileExtractor(), new RawImgFileExtractor(), new DefaultFileExtractor()};
-            
-            IEnumerable<string> extractedFiles = extractor.Extract(file, extractors).Select(x => x.LocalPath);
+            using var tmodFile = typeof(ExtractingTest).Assembly.GetManifestResourceStream("TML.Tests.GamerMod.tmod")!;
+            var file = TModFileSerializer.Deserialize(tmodFile);
+            IFileExtractor[] extractors = {new InfoFileExtractor(), new RawImgFileExtractor(), new RawByteFileExtractor()};
+
+            IEnumerable<string> extractedFiles = TModFileExtractor.Extract(file, 8, extractors).Select(x => x.Path);
             CollectionAssert.AreEquivalent(extractedFiles, Files);
         }
 
         [Test]
         // This is not a unit test.
         public static void ThisTestSucks() {
-            using Stream tmodFile = typeof(ExtractingTest).Assembly.GetManifestResourceStream("TML.Tests.GamerMod.tmod")!;
-            IModFileReader reader = new ModFileReader();
-            IModFileExtractor extractor = new ModFileExtractor();
-            IModFile file = reader.Read(tmodFile);
-            IFileExtractor[] extractors = {new InfoFileExtractor(), new RawImgFileExtractor(), new DefaultFileExtractor()};
+            using var tmodFile = typeof(ExtractingTest).Assembly.GetManifestResourceStream("TML.Tests.GamerMod.tmod")!;
+            var file = TModFileSerializer.Deserialize(tmodFile);
+            IFileExtractor[] extractors = {new InfoFileExtractor(), new RawImgFileExtractor(), new RawByteFileExtractor()};
 
-            Directory.Delete("GamerMod", true);
-            
-            IEnumerable<IExtractedModFile> extractedFiles = extractor.Extract(file, extractors);
-            foreach (IExtractedModFile extractedFile in extractedFiles) {
-                string path = Path.Combine("GamerMod", extractedFile.LocalPath);
+            if (Directory.Exists("GamerMod")) Directory.Delete("GamerMod", true);
+
+            IEnumerable<TModFileData> extractedFiles = TModFileExtractor.Extract(file, 8, extractors);
+            foreach (var extractedFile in extractedFiles) {
+                string path = Path.Combine("GamerMod", extractedFile.Path);
                 Directory.CreateDirectory(Path.GetDirectoryName(path)!);
                 File.WriteAllBytes(path, extractedFile.Data);
             }
